Fix =~ being called on a Time object
It seems that this argument is sometimes a time string and sometimes a
Time object. I expect this used to be mediated by some implicit type
conversion that has gone away in some newer version of Ruby, but
honestly I don't have a complete explanation. The fix seems harmless
enough.