Specify permitted_classes in YAML.load
It sounds like in response to CVE-2022-32224 the default YAML loader was
changed to safe_load, which constrains which classes can be deserialized
in the given data. This list of classes was sufficient for me to get a
poll up and running, though I'm not confident it's all the classes
needed for anything.